make uniqueness- and completeness checks pluggable
refs: #26139

There are different scenarios where the ContactInfoValidator is used.
Besides standard registrations and external users (who defined the initial
ruleset) there are also internal users (with additional completeness
rules) and registrations of other logistics applications who also define
additional required fields. So, completeness- and uniqueness-checks are
now extracted to an interface which can be provided by the user code.
The contargo-types library provides some implementations that should
cover the common use cases.
