+++ src/main/java/net/contargo/types/contactinfo/ContactInformationSegment.java
@@ -1,0 +1,8 @@
+package net.contargo.types.contactinfo;
+
+public enum ContactInformationSegment {
+
+    EMAIL,
+    MOBILE,
+    PHONE
+}
