+++ src/main/java/net/contargo/types/contactinfo/validation/CompletenessValidator.java
@@ -1,0 +1,11 @@
+package net.contargo.types.contactinfo.validation;
+
+import net.contargo.types.contactinfo.ContactInformation;
+
+import java.util.List;
+
+
+public interface CompletenessValidator {
+
+    List<ValidationResult> checkCompleteness(ContactInformation contactInformation);
+}
